Fix selecting current organization
Setting the selected organization in the componentDidMount() lifecycle
method in `OrganizationDetail` and `OrganizationSettings` isn't
enough, since the components get recycled and used for the rendering
of other organizations as well. So we would need to set the selected
organization in the `componentDidUpdate()` method as well.
However, this is not a good idea, as it results in an endless update
loop.

To fix this issue, the selected organization is now set during the
`onLoad` hook of the routes for the initial load and in the
`selectOrganization` history listener for the "update case".