Update counters for landings, flight and engine hours on aircraft
Before, only the number of flights was counted. Now, we sum up the
number of landings and the flight and engine hours as well.

The counters are updated whenever the data on a flight is modified
or a flight is created or deleted.
A complete recalculation of the counters is triggered whenever the
counter object of the aircraft or one of its properties gets removed.