Deploy `develop` branch to staging system
Until now, all changes were committed directly into the master
branch and deployed to the staging and production system right away.
As the first users have started to work with the production system now,
we develop the coming versions in the `develop` branch and merge the
changes into the master branch once they're ready to be released.