Fix getting destination aerodrome of last flight when creating new one
`getDestinationAerodrome` still assumed that `flight.destinationAerodrome`
is an instance of `DocumentReference` and that the actual data has to be
resolved using the `get` function.
However, newer flights don't just have a reference to the aerodrome,
but also have the actual data of the aerodrome stored in the
`destinationAerodrome` property. So in this case, we don't have to
resolve a reference but can return the data right away.