Make sure that no flight can be created with date prior to last flight
If the same date as the date of the last flight was entered with a
block off time before the block on time of the last flight, this was
already recognized as invalid.
However, if the entered date was before the last flight, the block off
time of the new record was compared with the date of today instead of
the actually entered date. This issue is fixed now.