Show message to user if has no access for organization (or not exists)
For this reason, fetching of organization has to be changed. We can't
watch them anymore, but will fetch them on login instead (this will
also be executed when the user reloads the browser window). We have to
do it this way, because we need to be able to distinguish if the
organizations have been loaded or not. If we watch them, we receive
updates continuously and we never reach a "final" state.