Fix trigger that updates the organizations of user when member changes
The condition `data[fieldName] == previousData[fieldName]` for the
immediate exit if the `user` reference points to the same user document
before and after the change didn't work correctly -> it always
returned false -> as a result, the organization was removed in the user
document way too often.

To fix this bug, we compare the id of the referenced user document
instead of the whole reference.