Keep orgs in loading state if not set on user and fetch on user update
The `orgs` field of the user will be undefined on the first login
after the new organization handling is released. There is a cloud
trigger that automatically collects the organizations of the user
when he logs in and sets them on the current user.

So, we'll just show the user the loading icon during the migration
of the organizations and fetch the organizations again when the
current user is updated.