Show troubles/observations section even if techlog not enabled
If the techlog is not enabled, there is only the nil/troubles radio
group in the flight creation dialog where the pilot has to choose one
option. If he selects "troubles", he has to enter a description (there
is no status field and there are no attachments). The text description
is written to the flight record without creating an additional techlog
entry (just like in the paper book).