Fix how the lock date is stored
If the Firestore timestamp is created from the date string '2020-11-05',
it is actually stored as '2020-11-05 00:00:00'.

This is not what we want, because then the comparison of the flight date
in the security rules does not work:
`aircraftDoc.data.settings.lockDate.toMillis() < res.data.blockOffTime.toMillis()`

If we set the 2020-11-05 as lock date, we want everything to be forbidden
*including* 2020-11-05.

Therefore, we have to store the date as `2020-11-06 00:00:00` to make
the security rules condition work.